Finished modular typography and spacing
Used island and push none to remove padding and create correct spacing
in footer.
Worked out spacing in header so that the title moves and aligns
properly and their is correct spacing around the ul’s. Also created
padding around the ul’s so that when clicked would be a larger square.
Fixed spacing around the images and divs